🔨 Working executable JAR that also serves the REST API
Now every module that uses the spring-conventions plugin also builds a separate -boot.jar.
Normally only gradle modules that use the `java` plugin, will be targeted by the springboot-gradle plugin, but it appears all our kotlin modules get targeted when they're combined with the springboot-gradle plugin.
So, as a "work-around" we re-enable the regular "jar" task that springboot-gradle disables, so we get a regular library that the other modules can depend on.
Another part of the work-around is to also suffix the created bootjar with -boot, so the regular jar doesn't get overridden with the bootjar. This we do by configuring the BootJar task with `classifier = "boot"`.
We also need to set the "mainClassName", otherwise the gradle build complains on the BootJar task execution, telling us there's not mainClassName configured.

Ideally we figure out WHY our kotlin modules are being recognized as `java` modules, or we figure out how we can disable the springboot gradle plugin execution, which we initially tried by typing `id("org.springframework.boot") apply false` but didn't seem to work.